mcp/requirements.txt
fastmcp
tavily-python
selenium
requests
langchain-google-genai
python-dotenv
langgraph
yfinance
alpha-vantage
finnhub-python
newsapi-python
